requirements.txt
# 视频处理核心库
opencv-python>=4.8.0
ffmpeg-python>=0.2.0
moviepy>=1.0.3

# CLI框架
click>=8.1.0

# 配置管理
python-dotenv>=1.0.0

# 测试框架
pytest>=7.4.0
pytest-cov>=4.1.0

# 进度条显示
tqdm>=4.66.0

# 类型检查（可选）
# mypy>=1.5.0
# types-python-dotenv>=1.0.0

